UPSTREAM: net: bridge: vlan: fix global vlan option range dumping
commit 6c4110d9f499e2170fbb36723b0a5f50a8116304 upstream.

When global vlan options are equal sequentially we compress them in a
range to save space and reduce processing time. In order to have the
proper range end id we need to update range_end if the options are equal
otherwise we get ranges with the same end vlan id as the start.

Fixes: 743a53d9636a ("net: bridge: vlan: add support for dumping global vlan options")
Change-Id: I244048954dd1c3d71c353c8c9237ac4988e0b1a4
Signed-off-by: Nikolay Aleksandrov <nikolay@nvidia.com>
Link: https://lore.kernel.org/r/20210810092139.11700-1-razor@blackwall.org
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 66157235c8d0fefc152c11c962792d98139deea5)
Signed-off-by: CloudFox-INC <CloudFox_INC@proton.me>